+++ src/animations.h
@@ -137,7 +137,7 @@ int32 set_model_animation(int32 animIdx, uint8 *animPtr, uint8 *bodyPtr, AnimTim
 /** Get entity anim index (This is taken from File3D entities)
 	@param animIdx Entity animation index
 	@param actorIdx Actor index */
-int32 get_body_anim_index(int32 animIdx, int16 actorIdx);
+int32 get_body_anim_index(int32 animIdx, int32 actorIdx);
 
 /** Stock animation - copy the next keyFrame from a different buffer
 	@param animPtr Animation pointer
@@ -157,11 +157,11 @@ int32 verify_anim_at_keyframe(int32 animPos, uint8 *animPtr, uint8 *bodyPtr, Ani
 	@param animType animation type
 	@param animExtra animation actions extra data
 	@param actorIdx actor index */
-int32 init_anim(int8 newAnim, int16 animType, uint8 animExtra, int16 actorIdx);
+int32 init_anim(int32 newAnim, int16 animType, uint8 animExtra, int32 actorIdx);
 
 /** Process acotr animation actions
 	@param actorIdx Actor index */
-void process_anim_actions(int16 actorIdx);
+void process_anim_actions(int32 actorIdx);
 
 /** Process main loop actor animations
 	@param actorIdx Actor index */
